Logical restructuring of rest senders
