+++ src/test/java/org/radarcns/producer/rest/RestSenderTest.java
@@ -16,25 +16,10 @@
 
 package org.radarcns.producer.rest;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.JsonNodeType;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.AbstractMap.SimpleImmutableEntry;
-import java.util.Arrays;
-import java.util.Map.Entry;
-import java.util.zip.GZIPInputStream;
 import okhttp3.Headers;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
@@ -50,8 +35,15 @@ import org.radarcns.data.SpecificRecordEncoder;
 import org.radarcns.key.MeasurementKey;
 import org.radarcns.phone.PhoneLight;
 import org.radarcns.producer.KafkaTopicSender;
-import org.radarcns.producer.SchemaRetriever;
 import org.radarcns.topic.AvroTopic;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Arrays;
+import java.util.zip.GZIPInputStream;
+
+import static org.junit.Assert.*;
+import static org.mockito.Mockito.*;
 
 public class RestSenderTest {
     private SchemaRetriever retriever;
@@ -80,9 +72,11 @@ public class RestSenderTest {
         Schema valueSchema = PhoneLight.getClassSchema();
         AvroTopic<MeasurementKey, PhoneLight> topic = new AvroTopic<>("test",
                 keySchema, valueSchema, MeasurementKey.class, PhoneLight.class);
-        sender.setHeaders(Arrays.<Entry<String, String>>asList(
-                new SimpleImmutableEntry<>("Cookie", "ab"),
-                new SimpleImmutableEntry<>("Cookie", "bc")));
+        Headers headers = new Headers.Builder()
+                .add("Cookie", "ab")
+                .add("Cookie", "bc")
+                .build();
+        sender.setHeaders(headers);
         KafkaTopicSender<MeasurementKey, PhoneLight> topicSender = sender.sender(topic);
 
         MeasurementKey key = new MeasurementKey("a", "b");
@@ -117,19 +111,9 @@ public class RestSenderTest {
         JsonNode records = body.get("records");
         assertEquals(JsonNodeType.ARRAY, records.getNodeType());
         assertEquals(1, records.size());
-        for (JsonNode child : records) {
-            JsonNode jsonKey = child.get("key");
-            assertEquals(JsonNodeType.OBJECT, jsonKey.getNodeType());
-            assertEquals("a", jsonKey.get("userId").asText());
-            assertEquals("b", jsonKey.get("sourceId").asText());
-            JsonNode jsonValue = child.get("value");
-            assertEquals(JsonNodeType.OBJECT, jsonValue.getNodeType());
-            assertEquals(0.1, jsonValue.get("time").asDouble(), 0);
-            assertEquals(0.2, jsonValue.get("timeReceived").asDouble(), 0);
-            assertEquals(0.3f, (float)jsonValue.get("light").asDouble(), 0);
-        }
-        Headers headers = request.getHeaders();
-        assertEquals(Arrays.asList("ab", "bc"), headers.values("Cookie"));
+        checkChildren(records);
+        Headers receivedHeaders = request.getHeaders();
+        assertEquals(Arrays.asList("ab", "bc"), receivedHeaders.values("Cookie"));
     }
 
     @Test
@@ -174,17 +158,7 @@ public class RestSenderTest {
         JsonNode records = body.get("records");
         assertEquals(JsonNodeType.ARRAY, records.getNodeType());
         assertEquals(2, records.size());
-        for (JsonNode child : records) {
-            JsonNode jsonKey = child.get("key");
-            assertEquals(JsonNodeType.OBJECT, jsonKey.getNodeType());
-            assertEquals("a", jsonKey.get("userId").asText());
-            assertEquals("b", jsonKey.get("sourceId").asText());
-            JsonNode jsonValue = child.get("value");
-            assertEquals(JsonNodeType.OBJECT, jsonValue.getNodeType());
-            assertEquals(0.1, jsonValue.get("time").asDouble(), 0);
-            assertEquals(0.2, jsonValue.get("timeReceived").asDouble(), 0);
-            assertEquals(0.3f, (float)jsonValue.get("light").asDouble(), 0);
-        }
+        checkChildren(records);
     }
 
     @Test
@@ -252,17 +226,21 @@ public class RestSenderTest {
             JsonNode records = body.get("records");
             assertEquals(JsonNodeType.ARRAY, records.getNodeType());
             assertEquals(1, records.size());
-            for (JsonNode child : records) {
-                JsonNode jsonKey = child.get("key");
-                assertEquals(JsonNodeType.OBJECT, jsonKey.getNodeType());
-                assertEquals("a", jsonKey.get("userId").asText());
-                assertEquals("b", jsonKey.get("sourceId").asText());
-                JsonNode jsonValue = child.get("value");
-                assertEquals(JsonNodeType.OBJECT, jsonValue.getNodeType());
-                assertEquals(0.1, jsonValue.get("time").asDouble(), 0);
-                assertEquals(0.2, jsonValue.get("timeReceived").asDouble(), 0);
-                assertEquals(0.3f, (float)jsonValue.get("light").asDouble(), 0);
-            }
+            checkChildren(records);
+        }
+    }
+
+    private static void checkChildren(JsonNode records) {
+        for (JsonNode child : records) {
+            JsonNode jsonKey = child.get("key");
+            assertEquals(JsonNodeType.OBJECT, jsonKey.getNodeType());
+            assertEquals("a", jsonKey.get("userId").asText());
+            assertEquals("b", jsonKey.get("sourceId").asText());
+            JsonNode jsonValue = child.get("value");
+            assertEquals(JsonNodeType.OBJECT, jsonValue.getNodeType());
+            assertEquals(0.1, jsonValue.get("time").asDouble(), 0);
+            assertEquals(0.2, jsonValue.get("timeReceived").asDouble(), 0);
+            assertEquals(0.3f, (float)jsonValue.get("light").asDouble(), 0);
         }
     }
 }
