+++ src/test/java/org/radarcns/util/serde/KafkaAvroSerializerTest.java
@@ -34,7 +34,7 @@ import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
 import org.junit.Test;
 import org.radarcns.key.MeasurementKey;
-import org.radarcns.producer.SchemaRetriever;
+import org.radarcns.producer.rest.SchemaRetriever;
 import org.radarcns.producer.rest.ParsedSchemaMetadata;
 
 public class KafkaAvroSerializerTest {
