+++ testing/src/main/java/org/radarcns/mock/MockProducer.java
@@ -46,9 +46,9 @@ import org.radarcns.mock.config.MockDataConfig;
 import org.radarcns.mock.data.MockCsvParser;
 import org.radarcns.mock.data.RecordGenerator;
 import org.radarcns.producer.KafkaSender;
-import org.radarcns.producer.SchemaRetriever;
+import org.radarcns.producer.rest.SchemaRetriever;
 import org.radarcns.producer.direct.DirectSender;
-import org.radarcns.producer.rest.BatchedKafkaSender;
+import org.radarcns.producer.BatchedKafkaSender;
 import org.radarcns.producer.rest.ConnectionState;
 import org.radarcns.producer.rest.ManagedConnectionPool;
 import org.radarcns.producer.rest.RestSender;
